+++ R/model-methods.R
@@ -96,8 +96,6 @@
 #' # model. In this case, there are no tools so an
 #' # error message will occur.
 #'
-#' # Draw a mermaid diagram of the transitions
-#' draw_mermaid(model_sirconn)
 queuing_on <- function(x) UseMethod("queuing_on")
 
 #' @export
@@ -421,48 +419,4 @@ clone_model <- function(model) {
     clone_model_cpp(model),
     class = class(model)
   )
-}
-
-#' @rdname epiworld-methods
-#' @export
-#' @inheritParams epiworld-model-diagram
-#' @details `draw_mermaid` generates a mermaid diagram of the model. The
-#' diagram is saved in the specified output file (or printed to the standard
-#' output if the filename is empty). See [draw_mermaid_from_data()].
-#' @return
-#' - The `draw_mermaid` returns the mermaid diagram as a string.
-#' @importFrom utils capture.output
-draw_mermaid <- function(
-  model,
-  output_file = "",
-  allow_self_transitions = FALSE
-) {
-  stopifnot_model(model)
-  stopifnot_string(output_file)
-  stopifnot_bool(allow_self_transitions)
-
-  if (output_file != "") {
-    draw_mermaid_cpp(
-      model,
-      output_file,
-      allow_self_transitions
-    )
-
-    message("Diagram written to ", output_file)
-
-    diagram <- readChar(output_file, file.info(output_file)$size)
-    return(diagram)
-  } else {
-    diagram <- capture.output(draw_mermaid_cpp(
-      model,
-      output_file,
-      allow_self_transitions
-    ))
-
-    return(
-      as_epiworld_diagram(
-        paste(diagram, collapse = "\n")
-      )
-    )
-  }
 }