+++ NEWS.md
@@ -8,6 +8,8 @@
 
 * The `ModelSEIRMixingQuarantine` model was not officially tracking hospitalizations. This means that using the saver to extract hospitalizations would return zero.
 
+* Using `globalevent_fun()` with `run_multiple(..., nthreads > 1)` was not compatible due to R not being thread safe. `run_multiple()` warns the user about this and automatically switches `nthreads` to be one.
+
 
 # epiworldR 0.15.0-0